requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
PyJWT
sqlparse
tzdata
